% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{analyseGeneChains}
\alias{analyseGeneChains}
\title{Identify structural rearrangement events from the source data.}
\usage{
analyseGeneChains(
  sourceData,
  cnvData = NULL,
  minMeanRating = 4.5,
  minSupportedReads = 3,
  minQuality = 70,
  clusteringAlgorithm = "ckmeans",
  maxK = 20,
  minNumTranslocations = 7,
  minNumChromosomes = 3,
  pathDepthLimit = NULL
)
}
\arguments{
\item{sourceData}{The structural variation dataset.}

\item{cnvData}{Optional CNV data used in visualisation.}

\item{minMeanRating}{The threshold for the minimum mean rating of identified paths. For example, setting `minMeanRating=4`,
the function only returns identified paths which have mean rating 4 and higher.}

\item{minSupportedReads}{The threshold for the minimum number of supporting reads for a translocation. Used in preprocessing, see `cleanSVDataset` function.}

\item{minQuality}{The threshold for the minimum quality of a translocation. Used in preprocessing, see also `cleanSVDataset` function.}

\item{clusteringAlgorithm}{The clustering algorithm used for clustering translocation breakpoints. One of `["ckmeans", "gmm", "affinity"]` values accepted. See `clusterData` function.}

\item{maxK}{The upper threshold for selecting the optimal number of clusters during breakpoint clustering. See `clusterData` function.}

\item{minNumTranslocations}{The threshold for the minimum number of translocations in a path. For example, setting `minNumTranslocations=10`,
the function only returns identified paths which contain 10 or more translocations.}

\item{minNumChromosomes}{The threshold for the minimum number of chromosomes involved in a path. For example, setting `minNumChromosomes=5`,
the function only returns identified paths which contain translocation between 5 or more chromosomes.}

\item{pathDepthLimit}{The upper boundary for the depth of a path, see `depthFirstSearch` function.}
}
\value{
A list containing the results of analysis:
\itemize{
\item `clusteredData`, the source dataset annotated with assigned cluster labels,
\item `graph`, a `Graph` type object generated from the source data,
\item `events`, a list of identified chains/paths which represent potential events of massive structural rearrangement. Each event contains:
\itemize{
\item `path`, ,
\item `totalRating`, ,
\item `meanRating`, ,
\item `maxRating`, ,
\item `numChromosomes`, ,
},
\item `mutations`, a concatenated dataset of translocations that are part of individual identified paths,
\item `allPaths`, a concatenated list of all paths, represented by origin and target breakpoint IDs. Used for visualisation.
}
}
\description{
The function takes in the source structural variation data, performs necessary preprocessing,
and using graph network search identifies potential massive structural rearrangement events
based on the defined criteria.
}
