% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{assignPathRating}
\alias{assignPathRating}
\title{Assign rating to all Edges in a path list.}
\usage{
assignPathRating(graph, path)
}
\arguments{
\item{graph}{a Graph object from which the path was created.}

\item{path}{the path to rank - a list of `Breakpoint_id`s.}
}
\value{
a total ranking of the path as one integer.
}
\description{
Assign rating to all Edges in a path list.
}
