% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{class}
\name{Breakpoint-class}
\alias{Breakpoint-class}
\title{Breakpoint class}
\description{
Breakpoint class
}
\section{Slots}{

\describe{
\item{\code{ID}}{unique identifier. Format is `[Chromosome]_[Cluster]_[Position]`.}

\item{\code{node_id}}{the node this breakpoint belongs to. See \link[chromolyse]{Node}.}

\item{\code{position}}{the position of breakpoint on the chromosome.}

\item{\code{gene}}{a string representing which gene the breakpoint is on.}

\item{\code{encode}}{the encode value of the breakpoint.}

\item{\code{region}}{the region value of the breakpoint. May also be called class in SV data.}
}}

