% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{cleanSVDataset}
\alias{cleanSVDataset}
\title{Filter a structural variation dataset}
\usage{
cleanSVDataset(dataset, minSupportedReads = 3, minQuality = 70)
}
\arguments{
\item{dataset}{the dataset to filter}

\item{minSupportedReads}{the threshold value for minimum number of supporting reads for a translocation. Should be provided in the `DP` column.}

\item{minQuality}{the threshold value for minimum quality of translocations. Should be provided in the `Qual` column.}
}
\value{
the filtered dataset only containing translocations with the required properties.
}
\description{
For the purposes of event identification, only certain structural variants are viable.
The function filters a dataset to contain only such variants.
}
