% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster SV data}
\usage{
cluster_data(
  dataset,
  algorithm = "ckmeans",
  grouping_column = "Chr.x",
  start_position_column = "Position.x",
  end_position_column = "Position.y"
)
}
\arguments{
\item{dataset}{The input data.}

\item{algorithm}{Algorithm used to cluster position values from the dataset.}

\item{grouping_column}{Name of the column which contains chromosome assignments. This column will be used for grouping.}

\item{start_position_column}{Name of the column which contains starting positions of translocations on the chromosome. These will be used for clustering.}

\item{end_position_column}{Name of the column which contains end positions of translocations on the chromosome. These will be used for clustering.}
}
\value{
original dataset with cluster labels in Cluster.x column
}
\description{
Function performs clustering of SV data chromosome-wise using selected algorithm.
}
