% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{containsEdge}
\alias{containsEdge}
\title{Check if `Edge` is in a list of edges.}
\usage{
containsEdge(edge_list, target_edge)
}
\arguments{
\item{edge_list}{the list of edges.}

\item{target_edge}{the `Edge` to search for.}
}
\value{
boolean whether the `Edge` was found.
}
\description{
Check if `Edge` is in a list of edges.
}
