% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{createDataframeFromEvent}
\alias{createDataframeFromEvent}
\title{Extract information from identified paths to a data table.}
\usage{
createDataframeFromEvent(graph, event)
}
\arguments{
\item{graph}{the graph used for event identification}

\item{event}{the event object = a list generated by the `\link[chromolyse]{analyseGeneChains}` function.}
}
\value{
a data table containing all translocations present in the event.
}
\description{
Extract information from identified paths to a data table.
}
