% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{createDataframeFromPath}
\alias{createDataframeFromPath}
\title{Transform a list of `Breakpoint` object IDs into a dataframe.}
\usage{
createDataframeFromPath(graph, path)
}
\arguments{
\item{graph}{a `Graph` object. Used to extract `Breakpoint` objects by their IDs.}

\item{path}{a list of `Breakpoint` object IDs.}
}
\value{
a data table with all translocations and their respective attributes.
}
\description{
Transform a list of `Breakpoint` object IDs into a dataframe.
}
