% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{createGraphFromDataframe}
\alias{createGraphFromDataframe}
\title{Read input data and create a `Graph` object from it.}
\usage{
createGraphFromDataframe(dataframe)
}
\arguments{
\item{dataframe}{the input dataframe of translocations.}
}
\value{
a `Graph` object
}
\description{
Read input data and create a `Graph` object from it.
}
