% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{depthFirstSearch}
\alias{depthFirstSearch}
\title{Perform Depth-First Search.}
\usage{
depthFirstSearch(
  graph,
  curr_node_id,
  curr_path = NULL,
  depth = 0,
  depthLimit = NULL
)
}
\arguments{
\item{graph}{`Graph` object to search.}

\item{curr_node_id}{`Node` ID to start from. The algorithm is recursive and will use this parameter to modify the current node.}

\item{curr_path}{a list of `Edge` objects representing traversal progress. Should stay NULL when calling.}

\item{depth}{the depth of traversal. Should stay 0 when calling.}

\item{depthLimit}{an upper boundary restricting the traversal depth.}
}
\value{
a list of `Edge` objects in order of traversal. This can be used to extract individual paths using the `extractAllPaths` function.
}
\description{
Perform Depth-First Search.
}
