% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.clusterAP}
\alias{.clusterAP}
\title{Perform clustering using Affinity Propagation algorithm.}
\usage{
.clusterAP(positions)
}
\arguments{
\item{positions}{input data to cluster. This should represent a subset of positions from one chromosome.}
}
\value{
cluster labels.
}
\description{
The algorithm is an implementation from the apcluster package, see \code{apcluster::\link[apcluster:apcluster]{apcluster}} for details.
}
\keyword{internal}
