% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.clusterGMM}
\alias{.clusterGMM}
\title{Perform clustering using the Expectation-Maximisation algorithm inside a Gaussian Mixture Model.}
\usage{
.clusterGMM(data, kUpperBound = 20)
}
\arguments{
\item{data}{explanation}

\item{kUpperBound}{explanation}
}
\value{
return value description
}
\description{
During the iteration of the EM algorithm, the log-likelihood of the data is calculated at each step.
The algorithm continues iterating until the change in log-likelihood between consecutive iterations is smaller than the tolerance (`tol`).
If the absolute difference in log-likelihood between two consecutive iterations falls below the tolerance value, the algorithm considers that it has converged and stops iterating.
}
\keyword{internal}
