% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.emCluster_k}
\alias{.emCluster_k}
\title{Perform clustering using Expectation-Maximisation algorithm}
\usage{
.emCluster_k(positions, k = 20, max_iterations = 100, tol = 1e-06)
}
\arguments{
\item{positions}{explanation}

\item{k_upper}{explanation}
}
\value{
return value description
}
\description{
During the iteration of the EM algorithm, the log-likelihood of the data is calculated at each step.
The algorithm continues iterating until the change in log-likelihood between consecutive iterations is smaller than the tolerance (`tol`).
If the absolute difference in log-likelihood between two consecutive iterations falls below the tolerance value, the algorithm considers that it has converged and stops iterating.
}
\keyword{internal}
