% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{class}
\name{Edge-class}
\alias{Edge-class}
\title{Edge class}
\description{
Represents an edge in the graph network. Also represents the translocation between two `Breakpoint` objects.
}
\section{Slots}{

\describe{
\item{\code{source_breakpoint_id}}{`Breakpoint` ID of the start breakpoint in the translocation.}

\item{\code{source_node_id}}{`Node` ID of the start breakpoint in the translocation.}

\item{\code{source_chromosome}}{chromosome which the start breakpoint belongs to.}

\item{\code{source_cluster}}{cluster which the start breakpoint has been assigned.}

\item{\code{source_position}}{start breakpoint position.}

\item{\code{destination_breakpoint_id}}{`Breakpoint` ID of the end breakpoint in the translocation.}

\item{\code{destination_node_id}}{ID of the end breakpoint in the translocation.}

\item{\code{destination_chromosome}}{chromosome which the end breakpoint belongs to.}

\item{\code{destination_cluster}}{cluster which the end breakpoint has been assigned.}

\item{\code{destination_position}}{end breakpoint position.}

\item{\code{edge_type}}{type of translocation: [`ITX`,`CTX`].}
}}

