% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{extractAllPaths}
\alias{extractAllPaths}
\title{Extract exact paths from a traversal list generated by Depth-First Search (see \link[chromolyse]{depthFirstSearch}).}
\usage{
extractAllPaths(graph, dfs_traversal)
}
\arguments{
\item{graph}{the graph on which the search has been performed.}

\item{dfs_traversal}{a sequence (list) of `Edge` objects returned by calling the `depthFirstSearch`.}
}
\value{
a list of all paths.
}
\description{
The function returns a list of individual paths. The list contains Breakpoint IDs, always beginning with the starting node and organized in pairs - odd indexes are starting positions of an edge, even positions are end positions of an edge.
The length of such list of paths should hence always be even.
}
