% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{fill_missing_cnv_segments}
\alias{fill_missing_cnv_segments}
\title{Fill missing CNV segments}
\usage{
fill_missing_cnv_segments(cnv_file)
}
\arguments{
\item{cnv_file}{a source file containing CNV information.}
}
\value{
a modified CNV file
}
\description{
Fill missing CNV segments
}
