% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{getAdjacencyMatrix}
\alias{getAdjacencyMatrix}
\title{Transform `Graph` nodes into an adjancency matrix.}
\usage{
getAdjacencyMatrix(graph)
}
\arguments{
\item{graph}{the `Graph` object from which the matrix will be created.}
}
\value{
the adjacency matrix. Each cell has the sum of all translocation ratings.
}
\description{
The function iterates over the list of `Edge` objects in `graph@edges` and accumulates the total rating of all translocations to a combination of start and destination `Node` IDs.
}
