% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{getBreakpointByID}
\alias{getBreakpointByID}
\title{Retrieve a `Breakpoint` object by its `ID`}
\usage{
getBreakpointByID(graph, breakpoint)
}
\arguments{
\item{graph}{the graph in which the `Breakpoint` should exist.}

\item{breakpoint}{the `Breakpoint` ID. Format is `[Chromosome]_[Cluster]_[Position]`.}
}
\value{
the matching `Breakpoint` object if exists else `NULL`.
}
\description{
Retrieve a `Breakpoint` object by its `ID`
}
