% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{getNumChromosomes}
\alias{getNumChromosomes}
\title{Extract number of chromosomes from a list of `Breakpoint` IDs representing a path.}
\usage{
getNumChromosomes(path)
}
\arguments{
\item{path}{a list of `Breakpoint` IDs. Likely a result of `extractAllPaths` function.}
}
\value{
an integer representing the number of chromosomes the `path` involves.
}
\description{
Extract number of chromosomes from a list of `Breakpoint` IDs representing a path.
}
