% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{class}
\name{Graph-class}
\alias{Graph-class}
\title{Graph class}
\description{
Represents the graph network as a list of nodes and a list of edges.
}
\section{Slots}{

\describe{
\item{\code{nodes}}{list of `Node` objects.}

\item{\code{edges}}{list of `Edge` objects.}
}}

