% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\docType{class}
\name{Node-class}
\alias{Node-class}
\title{Node class}
\description{
Represents a node in the graph network.
}
\section{Slots}{

\describe{
\item{\code{ID}}{unique identifier. Format is `[chromosome]_[cluster]`.}

\item{\code{chromosome}}{which chromosome the node is on.}

\item{\code{cluster}}{the cluster label assigned.}

\item{\code{breakpoints}}{a list of `Breakpoints` belonging to this node.}

\item{\code{neighbours}}{a list of `Breakpoints` that are the destination breakpoint in translocations which contain this node's translocations as start breakpoints.}
}}

