% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{preprocess_raw_CNV}
\alias{preprocess_raw_CNV}
\title{Title}
\usage{
preprocess_raw_CNV(dataset)
}
\arguments{
\item{dataset}{The input dataset containing CNV information}
}
\value{
Modified dataset
}
\description{
Title
}
