% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{preprocessRawBattenbergFile}
\alias{preprocessRawBattenbergFile}
\title{Transform the Battenberg standard format file into the required format.}
\usage{
preprocessRawBattenbergFile(dataset)
}
\arguments{
\item{dataset}{The input Battenberg format dataset containing CNV information.}
}
\value{
Modified dataset.
}
\description{
The Battenberg file format should contain these columns:
- `chr`
- `startpos`
- `endpos`
- `nMaj1_A`
- `nMin1_A`
- `frac1_A`
- `nMaj2_A`
- `nMin2_A`
- `frac2_A`
}
\details{
This function transforms the dataset by only selecting columns needed for
CNV purposes (mainly visualisation):
- `chromosome`, specifies which chromosome is concerned,
- `start_position`, the starting position of an affected region,
- `end_position`, the end position of an affected region,
- `event`, a numeric value (0,5) specifying the CNV event.
}
