% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{preprocessRawNexusFile}
\alias{preprocessRawNexusFile}
\title{Transform the Nexus standard format file into the required format.}
\usage{
preprocessRawNexusFile(dataset)
}
\arguments{
\item{dataset}{The input Nexus format dataset containing CNV information.}
}
\value{
Modified dataset.
}
\description{
The Nexus file format should contain these columns:
- `Chromosome.Region`
- `Event``
- `Length`
- `Cytoband`
- `X..of.CNV.Overlap`
- `Probe.Median`
- `X..Heterozygous`
- `Probes`
- `Count.of.Gene.Symbols`
}
\details{
This function transforms the dataset by only selecting columns needed for
CNV purposes (mainly visualisation) and using appropriate names:
- `chromosome`, specifies which chromosome is concerned,
- `start_position`, the starting position of an affected region,
- `end_position`, the end position of an affected region,
- `event`, a numeric value (0,5) specifying the CNV event.
}
