% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{rateEdge}
\alias{rateEdge}
\title{Assing rating to an edge.}
\usage{
rateEdge(graph, sourceBreakpointId, destinationBreakpointId)
}
\arguments{
\item{graph}{The graph network object `Graph` which should contain the edge.}

\item{sourceBreakpointId}{The string identificator of translocation origin breakpoint.}

\item{destinationBreakpointId}{The string identificator of translocation target breakpoint.}
}
\value{
The rating of the edge as a floating point value.
}
\description{
Function assigns rating to an arbitrary edge in a graph network generated from structural variation data.
The edge represents a translocation, defined by origin and target breakpoints and their properties.
}
