% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{readExcel}
\alias{readExcel}
\title{Read an excel file}
\usage{
readExcel(filepath, sheetname = 2)
}
\arguments{
\item{filepath}{A path to the excel file.}

\item{sheetname}{A value to specify which sheet will be read.}
}
\value{
table with the contents of the excel file.
}
\description{
Function is a wrapper for readxl::read_excel function (see\code{readxl::\link[readxl:read_excel]{read_excel function}} for details).
}
