% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{searchGraph}
\alias{searchGraph}
\title{Perform Depth-first search from all nodes in a graph}
\usage{
searchGraph(graph)
}
\arguments{
\item{graph}{the graph to search.}
}
\value{
a list of "traversals", a sequence of nodes visited in a list, for all nodes
}
\description{
Perform Depth-first search from all nodes in a graph
}
