% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{assignPathRating}
\alias{assignPathRating}
\title{Assign rating to a path (list of graph breakpoints).}
\usage{
assignPathRating(graph, path)
}
\arguments{
\item{graph}{A `Graph` object from which the path was created.}

\item{path}{The path to rank - a list of `Breakpoint_ID`s.}
}
\value{
The total ranking of the path as as floating point number.
}
\description{
Assign rating to a path (list of graph breakpoints).
}
