% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{cluster_data}
\alias{cluster_data}
\title{Cluster SV data}
\usage{
cluster_data(
  dataset,
  algorithm = "ckmeans",
  grouping_column = "Chr.x",
  position_column = "Position.x"
)
}
\arguments{
\item{dataset}{The input data.}

\item{algorithm}{Algorithm used to cluster position values from the dataset.}

\item{grouping_column}{Name of the column which contains chromosome assignments. This column will be used for grouping.}

\item{position_column}{Name of the column which contains positions on the chromosome. This will be used for clustering directly.}
}
\value{
original dataset with cluster labels in Cluster.x column
}
\description{
Function performs clustering of SV data chromosome-wise using selected algorithm.
}
\examples{
temp1 <- <function_name>(50);
temp2 <- <function_name>( c(50, 63, 23) );
}
