% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{clusterData}
\alias{clusterData}
\title{Cluster SV data}
\usage{
clusterData(dataset, algorithm = "ckmeans", maxK = 20)
}
\arguments{
\item{dataset}{The input data.}

\item{algorithm}{Algorithm used to cluster position values from the dataset.}

\item{maxK}{The maximum number of clusters allowed for clustering. User definable, default set to 20.}
}
\value{
original dataset with cluster labels in Cluster.x column for clustered Chr.x positions and cluster labels in Cluster.y for Chr.y positions
}
\description{
Function performs clustering of SV data chromosome-wise using selected algorithm.
}
