% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.clusterAP}
\alias{.clusterAP}
\title{Perform clustering using Affinity Propagation algorithm.}
\usage{
.clusterAP(data)
}
\arguments{
\item{data}{}
}
\value{
f
}
\description{
Perform clustering using Affinity Propagation algorithm.
}
