% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.clusterCkmeans}
\alias{.clusterCkmeans}
\title{Perform clustering using Ckmeans 1D algorithm}
\usage{
.clusterCkmeans(positions, kUpperBound = 20)
}
\arguments{
\item{positions}{input data to cluster. This should represent a subset of positions from one chromosome.}

\item{kUpperBound}{the upper limit for the K value (target number of clusters).}
}
\value{
cluster labels.
}
\description{
Perform clustering using Ckmeans 1D algorithm
}
\keyword{internal}
