% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.clusterCkmeans}
\alias{.clusterCkmeans}
\title{Perform clustering using Ckmeans 1D algorithm}
\usage{
.clusterCkmeans(positions, kUpperBound = 20)
}
\arguments{
\item{positions}{explanation}

\item{kUpperBound}{explanation}
}
\value{
return value description
}
\description{
Perform clustering using Ckmeans 1D algorithm
}
\keyword{internal}
