% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{.clusterGMM}
\alias{.clusterGMM}
\title{Perform clustering using the Expectation-Maximisation algorithm using a Gaussian Mixture Model.}
\usage{
.clusterGMM(positions, kUpperBound = 20)
}
\arguments{
\item{positions}{input data to cluster. This should represent a subset of positions from one chromosome.}

\item{kUpperBound}{the upper limit for the K value (target number of clusters).}
}
\value{
cluster labels.
}
\description{
GMM used from the Mclust library, see \code{mclust::\link[mclust:Mclust]{Mclust GMM}} for details.
}
\keyword{internal}
