% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rating.R
\name{extractAllPaths}
\alias{extractAllPaths}
\title{Extract exact paths from a traversal list generated by Depth-first search.}
\usage{
extractAllPaths(graph, dfs_traversal)
}
\arguments{
\item{graph}{the graph on which the search has been performed.}

\item{dfs_traversal}{a sequence of nodes visited in a list.}
}
\value{
a list of all paths.
The list contains Breakpoint IDs, always beginning with the starting node and organized in pairs - odd indexes are starting positions of an edge, even positions are end positions of an edge.
The length of such list of paths should hence always be even.
}
\description{
Extract exact paths from a traversal list generated by Depth-first search.
}
