% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{visualise_circos}
\alias{visualise_circos}
\title{Create a Circos visualisation}
\usage{
visualise_circos(
  sv_data,
  cnv_data = NULL,
  chromosome_selection = NULL,
  sv_focus = "all",
  cnv_track_style = "full"
)
}
\arguments{
\item{sv_data}{somatic variants (SV) source data file. Format of the data should be:}

\item{cnv_data}{An optional data source containing CNV information. Used to create a separate CNV track.}

\item{chromosome_selection}{an optional vector containing a selection of chromosomes to be included in the visualisation.}

\item{sv_focus}{the subset of translocations to focus on: "itx", "ctx" or "all".}

\item{cnv_track_style}{the type of CNV track visualisation: "full" is default and will display the CNV baseline event (Allelic imbalance, shown in black). Use "trim" to not include this baseline.}
}
\description{
Create a Circos visualisation
}
