% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{visualiseCircos}
\alias{visualiseCircos}
\title{Create a Circos visualisation}
\usage{
visualiseCircos(
  sv_data,
  cnv_data = NULL,
  chromosome_selection = NULL,
  sv_focus = "all",
  cnv_track_style = "full",
  link_color_cluster = "x"
)
}
\arguments{
\item{sv_data}{somatic variants (SV) source data file. Format of the data should be:
\itemize{
\item `Chr.x`, start breakpoint chromosome,
\item `Cluster.x`, start breakpoint cluster (generated by `cluster_data` function),
\item `Position.x`, start breakpoint position,
\item `Chr.y`, end breakpoint chromosome,
\item `Cluster.y`, end breakpoint cluster (generated by `cluster_data` function),
\item `Position.y`, end breakpoint position,
\item `Class.x`, start breakpoint class (intergenic, intronic, exonic,...),
\item `Gene.x`, start breakpoint gene,
\item `Encode.x`, start breakpoint encode (Promotor, Enhancer, Heterochrom,...),
\item `Class.y`, end breakpoint class (intergenic, intronic, exonic,...),
\item `Gene.y`, end breakpoint gene,
\item `Encode.y`, end breakpoint encode (Promotor, Enhancer, Heterochrom,...),

}}

\item{cnv_data}{An optional data source containing CNV information. Used to create a separate CNV track.
This should follow the required format (see `preprocess_raw_nexus_file` function).}

\item{chromosome_selection}{an optional vector containing a selection of chromosomes to be included in the visualisation.}

\item{sv_focus}{the subset of translocations to focus on: {`"itx"`, `"ctx"`, `"all"`}.}

\item{cnv_track_style}{the type of CNV track visualisation: `cnv_track_style = "full"` is default and will display the CNV baseline event (Allelic imbalance, shown in black).
Use `cnv_track_style = "trim"` to not include this baseline and only show Copy number gains and losses (red and blue segments).}
}
\description{
Create a Circos visualisation
}
